+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-# Your Minecraft, Your Way
-
 ## Changes:
-### Adds a toggle `settings.use-old-file-system`, It defaults to false however this is the system that was before the per-file change.
+* Updated to 1.20.6
+* Added folia support
+
+## Removal:
+* Removed the metrics toggle from `config.yml`
+    * To disable `metrics`, You can go in the `bStats` folder!
 
 ## Other:
 * [Feature Requests](https://github.com/Crazy-Crew/CrazyVouchers/discussions/categories/features)